src/main/java/com/matos/picpay/payments/notification/NotificationProducer.java
package com.matos.picpay.payments.notification;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.kafka.core.KafkaTemplate;
import org.springframework.stereotype.Service;

import com.matos.picpay.payments.transaction.Transaction;

@Service
public class NotificationProducer {
    private final KafkaTemplate<String, Transaction> kafkaTemplate;
    private static final Logger log = LoggerFactory.getLogger(NotificationProducer.class);

    public NotificationProducer(KafkaTemplate<String, Transaction> kafkaTemplate) {
        this.kafkaTemplate = kafkaTemplate;
    }

    public void sendNotification(Transaction transaction) {
        log.info("Transaction notifiyng: {}", transaction);
        kafkaTemplate.send("transaction-notification", transaction);
    }

}
